监听器bug修复 监听器增加主动开启/关闭功能，通过notify和wait实现

+++ src/main/java/com/zfg/learn/dao/SubscriptionMapper.java
@@ -25,7 +25,7 @@ public interface SubscriptionMapper {
      * @param fid
      * @return
      */
-    List<Long> selectFollowerByFid(@Param("fid") Integer fid);
+    List<Long> selectFollowerByFid(@Param("fid") Long fid);
 
     /**
      * 自定义条件查询 不建议经常使用
